+++ packages/@dicebear/collection/src/index.ts
@@ -13,3 +13,4 @@ export * as croodlesNeutral from '@dicebear/croodles-neutral';
 export * as pixelArt from '@dicebear/pixel-art';
 export * as pixelArtNeutral from '@dicebear/pixel-art-neutral';
 export * as openPeeps from '@dicebear/open-peeps';
+export * as personas from '@dicebear/personas';