+++ packages/@dicebear/personas/src/colors/background.ts
@@ -1,0 +1,14 @@
+import type { ColorGroup } from "../static-types";
+
+export const background: ColorGroup = {
+  'blue': 'rgba(147, 167, 255, 1)',
+  'turquoise': 'rgba(137, 230, 228, 1)',
+  'purple': 'rgba(179, 121, 247, 1)',
+  'green': 'rgba(169, 231, 117, 1)',
+  'red': 'rgba(255, 102, 116, 1)',
+  'yellow': 'rgba(255, 204, 101, 1)',
+  'pink': 'rgba(255, 122, 154, 1)',
+  'dark': 'rgba(7, 11, 40, 1)',
+  'grey': 'rgba(248, 251, 255, 1)',
+  'white': 'rgba(255, 255, 255, 1)',
+}