+++ packages/@dicebear/personas/src/colors/clothing.ts
@@ -1,0 +1,11 @@
+import type { ColorGroup } from "../static-types";
+
+export const clothing: ColorGroup = {
+  'blue': 'rgba(69, 109, 255, 1)',
+  'turquoise': 'rgba(84, 215, 199, 1)',
+  'purple': 'rgba(117, 85, 202, 1)',
+  'green': 'rgba(109, 187, 88, 1)',
+  'red': 'rgba(226, 69, 83, 1)',
+  'yellow': 'rgba(243, 182, 58, 1)',
+  'pink': 'rgba(245, 93, 129, 1)',
+}