+++ packages/@dicebear/personas/src/colors/index.ts
@@ -1,0 +1,4 @@
+export { background } from './background';
+export { hair } from './hair';
+export { clothing } from './clothing';
+export { skin } from './skin';