+++ packages/@dicebear/personas/src/components/index.ts
@@ -1,0 +1,6 @@
+export { eyes } from './eyes';
+export { hair } from './hair';
+export { body } from './body';
+export { mouth } from './mouth';
+export { nose } from './nose';
+export { facialHair } from './facialHair';